Move to lib/itertools
